JWT middleware added to routes

+++ back/routes/api.php
@@ -19,16 +19,16 @@ use App\Http\Controllers\AuthController;
 //R de CRUD
 /* Route::get('/', [UserController::class, 'index'])->name('usersApi');
 Route::get('/home', [UserController::class, 'index'])->name('usersApi'); */
-Route::get('/showUser/{id}', [UserController::class, 'show'])->name('showUserApi');
+Route::get('/showUser/{id}', [UserController::class, 'show'])->middleware('auth', 'JwtAuth')->name('showUserApi');
 
 //D de CRUD
-Route::delete('/deleteUser/{id}',[UserController::class,'destroy'])->name('destroyUserApi');
+Route::delete('/deleteUser/{id}',[UserController::class,'destroy'])->middleware('auth', 'JwtAuth')->name('destroyUserApi');
 
 //C de CRUD
 /* Route::post('/createUser',[UserController::class,'store'])->name('createUserApi'); */
 
 //U de CRUD
-Route::put('/updateUser/{id}', [UserController::class,'update'])->name('updateUserApi');
+Route::put('/updateUser/{id}', [UserController::class,'update'])->middleware('auth', 'JwtAuth')->name('updateUserApi');
 
 
 //Auth group
